refactor(shared): move divider to shared
